refactor : AuthLayout 공통 레이아웃 제작
