+++ src/pages/LoginPage.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
-import LoginForm from '../components/login/LoginForm';
-import LinkToSignupPage from '../components/login/LinkToSignupPage';
+import AuthLayout from '@/components/common/auth/AuthLayout';
+// import LoginForm from '../components/login/LoginForm';
+// import LinkToSignupPage from '../components/login/LinkToSignupPage';
 
 const LoginPage = () => (
-  <div>
-    <img
-      src="/assets/logo.svg"
-      alt="logo"
-      className="ml-auto mr-auto mt-[104px] mb-[40px]"
-    />
-    <LoginForm />
-    <LinkToSignupPage />
-  </div>
+  <AuthLayout>
+    <div>1</div>
+    {/* <LoginForm />
+    <LinkToSignupPage /> */}
+  </AuthLayout>
 );
 
 export default LoginPage;
