+++ src/styles/customScrollbar.css
@@ -3,7 +3,7 @@
 }
 
 .custom-scrollbar::-webkit-scrollbar-track {
-  background-color: #c7e5c7; /* 밝은 녹색 */
+  background-color: #c7e5c7;
   border-radius: 10px;
 }
